+++ src/cevf_ev.h
@@ -55,7 +55,7 @@ struct thstat_s;
 
 ev_th_id_t ev_add_th(struct thprop_s pd);
 int ev_rm_th(ev_th_id_t id);
-void ev_init(struct thpillar_s pillars[], uint8_t pillars_len);
+void ev_init(struct thpillar_s pillars[], size_t pillars_len);
 void ev_deinit(void);
 struct thstat_s *ev_run(struct thprop_s props[], uint8_t props_len);
 int ev_join(struct thstat_s *thstat);
